@@ -28,18 +28,18 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/animations": "^7.1.4",
-    "@angular/cli": "^7.1.4",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/forms": "^7.1.4",
+    "@angular/animations": "^8.2.14",
+    "@angular/cli": "^8.3.25",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
+    "@angular/language-service": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -58,11 +58,11 @@
     "moment": "^2.19.3",
     "node-sass": "4.13.0",
     "protractor": "~5.1.2",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.4",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~3.1.6",
-    "zone.js": "^0.8.26"
+    "typescript": "~3.5.3",
+    "zone.js": "^0.9.1"
   },
   "dependencies": {
     "tslib": "^1.9.0"
